+++ packages/core/select/src/components/input-options.js
@@ -113,7 +113,14 @@ class ControlOption extends Component<OptionProps, OptionState> {
   onMouseUp = () => this.setState({ isActive: false });
   onMouseLeave = () => this.setState({ isActive: false });
   render() {
-    const { getStyles, Icon, children, innerProps, ...rest } = this.props;
+    const {
+      getStyles,
+      Icon,
+      children,
+      innerProps,
+      innerRef,
+      ...rest
+    } = this.props;
 
     // prop assignment
     const props: InnerProps = {
@@ -126,7 +133,7 @@ class ControlOption extends Component<OptionProps, OptionState> {
     const [styles, classes] = getPrimitiveStyles({ getStyles, ...rest });
 
     return (
-      <div css={styles} className={classes} {...props}>
+      <div css={styles} className={classes} ref={innerRef} {...props}>
         <div css={iconWrapperCSS()}>
           <Icon
             primaryColor={getPrimaryColor({ ...this.props, ...this.state })}