+++ packages/editor/editor-core/src/__tests__/unit/plugins/text-formatting/input-rules.ts
@@ -687,6 +687,29 @@ describe('text-formatting input rules', () => {
         doc(p('testing – testing → testing ', code('code'))),
       );
     });
+
+    it('should not apply when prefixed by text', () => {
+      const { editorView, sel } = editor(doc(p('words`code{<>}')));
+      insertText(editorView, '`', sel);
+
+      expect(editorView.state.doc).toEqualDocument(doc(p('words`code`{<>}')));
+    });
+
+    it('should not apply when prefixed by text and parentheses', () => {
+      const { editorView, sel } = editor(doc(p('words(`code{<>}')));
+      insertText(editorView, '`', sel);
+
+      expect(editorView.state.doc).toEqualDocument(doc(p('words(`code`{<>}')));
+    });
+
+    it('should apply when prefixed by a space and parentheses', () => {
+      const { editorView, sel } = editor(doc(p('words (`code{<>}')));
+      insertText(editorView, '`', sel);
+
+      expect(editorView.state.doc).toEqualDocument(
+        doc(p('words (', code('code'), '{<>}')),
+      );
+    });
   });
 
   describe('nested rules', () => {