+++ packages/editor/editor-core/src/plugins/analytics/types/enums.ts
@@ -14,12 +14,14 @@ export const enum ACTION {
   STARTED = 'started',
   STOPPED = 'stopped',
   SUBSTITUTED = 'autoSubstituted',
+  VISITED = 'visited',
 }
 
 export const enum INPUT_METHOD {
   ASCII = 'ascii',
   AUTO = 'auto',
   AUTO_DETECT = 'autoDetect',
+  CARD = 'card',
   CLIPBOARD = 'clipboard',
   DRAG_AND_DROP = 'dragAndDrop',
   EXTERNAL = 'external',
@@ -34,6 +36,8 @@ export const enum INPUT_METHOD {
   SHORTCUT = 'shortcut',
   TOOLBAR = 'toolbar',
   TYPEAHEAD = 'typeAhead',
+  CONTEXT_MENU = 'contextMenu',
+  BUTTON = 'button',
 }
 
 export const enum ACTION_SUBJECT {
@@ -43,8 +47,10 @@ export const enum ACTION_SUBJECT {
   MEDIA = 'media',
   PANEL = 'panel',
   PICKER = 'picker',
+  SMART_LINK = 'smartLink',
   TEXT = 'text',
   TYPEAHEAD = 'typeAhead',
+  TABLE = 'table',
 }
 
 export const enum ACTION_SUBJECT_ID {
@@ -53,6 +59,8 @@ export const enum ACTION_SUBJECT_ID {
   BUTTON_HELP = 'helpButton',
   BUTTON_FEEDBACK = 'feedbackButton',
   CANCEL = 'cancel',
+  CARD_INLINE = 'inlineCard',
+  CARD_BLOCK = 'blockCard',
   CODE_BLOCK = 'codeBlock',
   DECISION = 'decision',
   DIVIDER = 'divider',