+++ packages/editor/editor-core/src/plugins/table/ui/LayoutButton/index.tsx
@@ -11,8 +11,8 @@ import CollapseIcon from '@atlaskit/icon/glyph/editor/collapse';
 import commonMessages from '../../../../messages';
 import ToolbarButton from '../../../../ui/ToolbarButton';
 import { TableCssClassName as ClassName } from '../../types';
-import { toggleTableLayout } from '../../actions';
 import { layoutButtonSize } from '../styles';
+import { toggleTableLayoutWithAnalytics } from '../../actions-with-analytics';
 
 export interface Props {
   editorView: EditorView;
@@ -96,7 +96,7 @@ class LayoutButton extends React.Component<Props & InjectedIntlProps, any> {
 
   private handleClick = () => {
     const { state, dispatch } = this.props.editorView;
-    toggleTableLayout(state, dispatch);
+    toggleTableLayoutWithAnalytics()(state, dispatch);
   };
 }