+++ packages/elements/elements-test-helpers/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/elements-test-helpers
 
+## 0.5.0
+- [minor] [b0210d7ccc](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/b0210d7ccc):
+
+  - reset jest modules before hydration
+
 ## 0.4.1
 - [patch] [1bcaa1b991](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1bcaa1b991):