+++ packages/navigation/atlassian-switcher/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/atlassian-switcher
 
+## 0.3.2
+- [patch] [aacc698f07](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/aacc698f07):
+
+  - Adds an analytics event to track atlassian switcher dissmisals using the triggerXFlow callback
+
 ## 0.3.1
 - [patch] [57f774683f](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/57f774683f):